readme

+++ Utils/Classifier_utils.py
@@ -131,8 +131,10 @@ def convert_features_to_tensors(features, batch_size):
         [f.input_ids for f in features], dtype=torch.long)
     all_input_mask = torch.tensor(
         [f.input_mask for f in features], dtype=torch.long)
-    all_segment_ids = torch.tensor([f.segment_ids for f in features])
-    all_label_ids = torch.tensor([f.label_id for f in features])
+    all_segment_ids = torch.tensor(
+        [f.segment_ids for f in features], dtype=torch.long)
+    all_label_ids = torch.tensor(
+        [f.label_id for f in features], dtype=torch.long)
 
     data = TensorDataset(all_input_ids, all_input_mask,
                          all_segment_ids, all_label_ids)
